requirements.txt
# ── Core AI / Agent Framework ─────────────────────────────
langgraph>=0.2.0
llama-index>=0.11.0
llama-index-retrievers-bm25>=0.3.0
llama-index-vector-stores-chroma>=0.3.0

# ── Vector Database ────────────────────────────────────────
chromadb>=0.5.0

# ── Embedding & Duplicate Checker ─────────────────────────
sentence-transformers>=3.0.0

# ── Reranker ──────────────────────────────────────────────
# cross-encoder/ms-marco-MiniLM-L-6-v2 diload via sentence-transformers
# tidak perlu package terpisah

# ── LLM Providers ─────────────────────────────────────────
google-generativeai>=0.8.0       # Gemini (dev)
groq>=0.11.0                     # Groq (dev)
anthropic>=0.40.0                # Claude (production)

# ── Schema & Validation ───────────────────────────────────
pydantic>=2.0.0

# ── Config & Environment ──────────────────────────────────
python-dotenv>=1.0.0
pyyaml>=6.0.0

# ── Backend API ───────────────────────────────────────────
fastapi>=0.115.0
uvicorn>=0.30.0

# ── Frontend UI ───────────────────────────────────────────
streamlit>=1.39.0

# ── Database ──────────────────────────────────────────────
# SQLite sudah built-in di Python, tidak perlu package tambahan

# ── Testing ───────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.24.0

# ── Utilities ─────────────────────────────────────────────
httpx>=0.27.0          # async HTTP client (untuk API internal)